refactor: move "<count> blocks deep" display and related logic to new TransactionConfirmations component
